shell attach thread in each callback & no detach
still need a better general way to attach the nng threads to the gc.
removed the ctor and dtor attachs because we don't want to call them
each time a requests comes in.
thread_dettach is not called, which could potentially cause issues if
nng respawns a thread?
But it should also not be called for each callback.